Выделяет comments в отдельный файл
