+++ js/main.js
@@ -2,3 +2,4 @@ import './util.js';
 import './data.js';
 import './thumbnails.js';
 import './fulls-photos.js';
+import './comments.js';
